refactor(merge): isolate single-file writers
