+++ src/search/src/datafusion/merge/mod.rs
@@ -16,11 +16,7 @@
 use std::sync::Arc;
 
 use arrow::array::RecordBatch;
-use config::{
-    FileFormat, get_config,
-    meta::stream::FileMeta,
-    utils::parquet::{VORTEX_FILE_META_KEY, encode_vortex_file_meta, new_parquet_writer},
-};
+use config::{get_config, meta::stream::FileMeta};
 use datafusion::{
     arrow::datatypes::Schema,
     catalog::TableProvider,
@@ -32,28 +28,16 @@ use parquet::{
     arrow::{AsyncArrowWriter, async_writer::AsyncFileWriter},
     file::metadata::KeyValue,
 };
-use vortex::{
-    VortexSessionDefault,
-    array::ArrayRef,
-    arrow::{FromArrowArray, FromArrowType},
-    dtype::DType,
-    file::VortexWriteOptions,
-    io::session::RuntimeSessionExt,
-    session::VortexSession,
-};
 
 use super::table_provider::uniontable::NewUnionTable;
-use crate::datafusion::{
-    exec::DataFusionContextBuilder,
-    sort_order::FileSortOrder,
-    vortex::{VORTEX_RUNTIME, vortex_write_strategy},
-};
+use crate::datafusion::{exec::DataFusionContextBuilder, sort_order::FileSortOrder};
 
 #[cfg(feature = "enterprise")]
 pub mod downsampling;
 mod merged_file;
 mod metrics;
 pub mod mode;
+mod single_file;
 
 pub use merged_file::{MergeResult, MergedFile};
 pub use mode::{MergeMode, MergeOutput};
@@ -64,29 +48,17 @@ pub async fn merge_parquet_files(
     schema: Arc<Schema>,
     tables: Vec<Arc<dyn TableProvider>>,
     bloom_filter_fields: &[String],
-    mut metadata: FileMeta,
+    metadata: FileMeta,
     mode: &MergeMode,
     output: MergeOutput,
 ) -> Result<MergeResult> {
     let start = std::time::Instant::now();
     let sql = mode.sql(&schema);
     log::debug!("merge_parquet_files [{mode}] sql: {sql}");
-    let (schema, mut rx, read_task) =
+    let (schema, rx, read_task) =
         run_merge_query(&sql, mode.output_sort_order(), schema, tables).await?;
 
     let files = match mode {
-        #[cfg(feature = "enterprise")]
-        MergeMode::Downsampling(_) => {
-            downsampling::write_files(
-                &schema,
-                bloom_filter_fields,
-                &metadata,
-                output.file_format,
-                rx,
-                read_task,
-            )
-            .await?
-        }
         MergeMode::MetricsIndexed => {
             metrics::write_files(
                 &schema,
@@ -99,32 +71,32 @@ pub async fn merge_parquet_files(
             )
             .await?
         }
-        _ => {
-            let buf = match output.file_format {
-                FileFormat::Parquet => {
-                    write_parquet(
-                        &schema,
-                        bloom_filter_fields,
-                        &metadata,
-                        output.parquet_compression,
-                        &mut rx,
-                        read_task,
-                    )
-                    .await?
-                }
-                FileFormat::Vortex => write_vortex(schema, &metadata, rx, read_task).await?,
-            };
-            metadata.compressed_size = buf.len() as i64;
-            vec![match mode {
-                MergeMode::MetricsHashSorted => MergedFile::MetricsHashSorted {
-                    data: buf,
-                    meta: metadata,
-                },
-                _ => MergedFile::Standard {
-                    data: buf,
-                    meta: metadata,
-                },
-            }]
+        MergeMode::Classic
+        | MergeMode::TraceTimeIndex
+        | MergeMode::FileList
+        | MergeMode::MetricsHashSorted => {
+            single_file::write(
+                schema,
+                bloom_filter_fields,
+                metadata,
+                mode,
+                output,
+                rx,
+                read_task,
+            )
+            .await?
+        }
+        #[cfg(feature = "enterprise")]
+        MergeMode::Downsampling(_) => {
+            downsampling::write_files(
+                &schema,
+                bloom_filter_fields,
+                &metadata,
+                output.file_format,
+                rx,
+                read_task,
+            )
+            .await?
         }
     };
 
@@ -203,86 +175,6 @@ async fn run_merge_query(
     Ok((schema, rx, read_task))
 }
 
-async fn write_parquet(
-    schema: &Arc<Schema>,
-    bloom_filter_fields: &[String],
-    metadata: &FileMeta,
-    compression: Option<&str>,
-    rx: &mut tokio::sync::mpsc::Receiver<RecordBatch>,
-    read_task: tokio::task::JoinHandle<Result<()>>,
-) -> Result<Vec<u8>> {
-    let mut buf = Vec::new();
-    let mut writer = new_parquet_writer(
-        &mut buf,
-        schema,
-        bloom_filter_fields,
-        metadata,
-        false,
-        compression,
-    );
-
-    let mut new_file_meta = metadata.clone();
-    new_file_meta.records = 0;
-    while let Some(batch) = rx.recv().await {
-        new_file_meta.records += batch.num_rows() as i64;
-        if let Err(e) = writer.write(&batch).await {
-            log::error!("merge_parquet_files write error: {e}");
-            return Err(e.into());
-        }
-    }
-
-    read_task
-        .await
-        .map_err(|e| DataFusionError::External(Box::new(e)))??;
-    append_metadata(&mut writer, &new_file_meta)?;
-    writer.close().await?;
-    Ok(buf)
-}
-
-async fn write_vortex(
-    schema: Arc<Schema>,
-    metadata: &FileMeta,
-    mut rx: tokio::sync::mpsc::Receiver<RecordBatch>,
-    read_task: tokio::task::JoinHandle<Result<()>>,
-) -> Result<Vec<u8>> {
-    // metadata segments belong to the write options, they can't be appended at
-    // close time like parquet's, so `records` may drift from the rows written
-    let file_meta = encode_vortex_file_meta(metadata);
-    let writer_task = VORTEX_RUNTIME.spawn_blocking(move || {
-        VORTEX_RUNTIME.block_on(async move {
-            let mut buf = Vec::new();
-            let session = VortexSession::default().with_tokio();
-            let dtype = DType::from_arrow(schema.as_ref());
-            let write_options = VortexWriteOptions::new(session.clone())
-                .with_strategy(vortex_write_strategy())
-                .with_metadata_segment(VORTEX_FILE_META_KEY, file_meta);
-            let mut writer = write_options.writer(&mut buf, dtype);
-
-            while let Some(batch) = rx.recv().await {
-                let array: ArrayRef = ArrayRef::from_arrow(batch, false).map_err(|e| {
-                    DataFusionError::Execution(format!(
-                        "Failed to convert arrow array to vortex array: {e}"
-                    ))
-                })?;
-                writer.push(array).await?;
-            }
-
-            writer.finish().await?;
-
-            Ok::<Vec<u8>, anyhow::Error>(buf)
-        })
-    });
-
-    read_task
-        .await
-        .map_err(|e| DataFusionError::External(Box::new(e)))??;
-
-    writer_task
-        .await
-        .map_err(|e| DataFusionError::Execution(format!("Vortex runtime task failed: {e}")))?
-        .map_err(|e| DataFusionError::Execution(format!("Failed to write vortex file: {e}")))
-}
-
 pub fn append_metadata<W: AsyncFileWriter>(
     writer: &mut AsyncArrowWriter<W>,
     file_meta: &FileMeta,
@@ -312,10 +204,9 @@ mod tests {
 
     use arrow::array::{Array, Int64Array, StringArray};
     use arrow_schema::{DataType, Field, Schema};
-    use config::{TIMESTAMP_COL_NAME, meta::stream::StreamType};
+    use config::{FileFormat, TIMESTAMP_COL_NAME, meta::stream::StreamType};
     use datafusion::datasource::MemTable;
     use parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
-    use vortex::file::OpenOptionsSessionExt;
 
     use super::*;
 
@@ -475,51 +366,5 @@ mod tests {
         // MAX(session_id) ignores NULLs: trace a keeps the one known session.
         assert_eq!(rows["a"], (100, Some("session-a".to_string()), 80, 140));
         assert_eq!(rows["b"], (120, None, 120, 125));
-    }
-
-    #[tokio::test]
-    async fn test_write_vortex_carries_file_meta() {
-        let schema = create_test_schema();
-        let batch = RecordBatch::try_new(
-            schema.clone(),
-            vec![
-                Arc::new(Int64Array::from(vec![100, 200, 300])),
-                Arc::new(StringArray::from(vec!["a", "b", "c"])),
-                Arc::new(Int64Array::from(vec![1, 2, 3])),
-            ],
-        )
-        .unwrap();
-
-        let metadata = FileMeta {
-            min_ts: 100,
-            max_ts: 300,
-            records: 3,
-            original_size: 1024,
-            ..Default::default()
-        };
-
-        let (tx, rx) = tokio::sync::mpsc::channel::<RecordBatch>(2);
-        tx.send(batch).await.unwrap();
-        drop(tx);
-        let read_task = tokio::task::spawn(async { Ok(()) });
-
-        let buf = write_vortex(schema, &metadata, rx, read_task)
-            .await
-            .unwrap();
-
-        let session = VortexSession::default().with_tokio();
-        let vxf = session
-            .open_options()
-            .include_metadata()
-            .open_buffer(vortex::buffer::Buffer::from(buf))
-            .unwrap();
-        let segment = vxf.metadata_segment(VORTEX_FILE_META_KEY).unwrap();
-        let file_meta: config::utils::json::Value =
-            config::utils::json::from_slice(segment.as_slice()).unwrap();
-        assert_eq!(file_meta["min_ts"], metadata.min_ts);
-        assert_eq!(file_meta["max_ts"], metadata.max_ts);
-        assert_eq!(file_meta["records"], metadata.records);
-        assert_eq!(file_meta["original_size"], metadata.original_size);
-        assert_eq!(vxf.row_count(), 3);
     }
 }
